Use long instead of int for nr_joints

+++ src/ubx/vis2b_osg_ubx.lua
@@ -11,7 +11,7 @@ return block
     },
 
     configurations = {
-        { name = "nr_joints", type_name = "int", min = 1, max = 1, doc = "Number of joints" },
+        { name = "nr_joints", type_name = "long", min = 1, max = 1, doc = "Number of joints" },
         { name = "mesh_base", type_name = "char", min = 0, doc = "Directory from which meshes should be loaded" },
         { name = "mesh_names", type_name = "char", min = 0, doc = "Colon-separated list of mesh file names" },
     },
